chore(solhint): wave 3 — named imports, inline disables, config rationale (warnings: 3,746 → 0) (#251)
This is the final wave of the solhint warning cleanup. The codebase
now lints clean (0 warnings, 0 errors).

## What this does

### 1. Convert global imports to named imports (94 imports, 52 files)
Replaces `import "./X.sol";` with `import { Symbol1, Symbol2 } from "./X.sol";`
where the imports are restricted to ONLY the symbols actually
referenced by the importing file (computed via AST analysis of both
the imported module's exports and the importer's identifier
references). This satisfies no-global-import without triggering
no-unused-import for symbols the importer does not actually use.

### 2. Fix one pre-existing import path bug
`contracts/BadgeQualificationRules.sol` imported
`./BadgeRegistry.sol` (non-existent) instead of
`./future/BadgeRegistry.sol` (where the file actually lives). This
was masked by the global-import format -- once converted to named
imports, the dangling reference surfaced. Fixed.

### 3. Drop one unused import
`contracts/future/BadgeManager.sol` imported the entire
`SharedInterfaces` module without using any symbol from it. Dropped.

### 4. Fix legacy/VaultInfrastructure.sol unused params (6 sites)
The disabled force-recovery stubs (`approveForceRecovery`,
`initiateForceRecovery`, `finalizeForceRecovery`,
`cancelForceRecovery`) are intentionally body-less revert stubs that
do not use their parameters. Renamed each parameter to `/* name */`
so the type stays in the signature but the unused-name warning
clears.

### 5. Inline-disable 6 SafeERC20-style low-level calls
The 6 `address.call()` invocations in SharedInterfaces SafeERC20,
RevenueSplitter._distribute, FeeDistributor._safeTransferOut, and
AdminMultiSig.executeProposal are deliberate -- they support
non-standard ERC-20s (USDT) or pass through a governance-approved
arbitrary call. Each got a
`// solhint-disable-next-line avoid-low-level-calls` annotation. The
rule itself stays `warn` so any NEW low-level call is flagged for
review.

### 6. .solhint.json: disable rules that conflict with project policy
Rules disabled fall into 5 categories, with full rationale documented
in SOLHINT.md:

  a. ABI-affecting style rules (renaming would break every external
     integration): func-name-mixedcase, var-name-mixedcase,
     const-name-snakecase, immutable-vars-naming,
     contract-name-capwords, event-name-capwords,
     interface-starts-with-i.

  b. Behavior- / boundary-changing gas rules (would silently change
     semantics or UX): gas-custom-errors, gas-indexed-events,
     gas-strict-inequalities, gas-small-strings, gas-struct-packing,
     gas-increment-by-one (22 leftover semantic-bearing cases),
     gas-calldata-parameters, reason-string.

  c. Idiomatic-empty-block false positives: no-empty-blocks
     (empty try/catch is the idiomatic Solidity pattern for
     best-effort external integration).

  d. Architectural-refactor rules requiring multi-week design work:
     function-max-lines, max-states-count, one-contract-per-file,
     no-complex-fallback.

  e. Convention rules: no-inline-assembly (used intentionally for
     EIP-712 caching and gas-critical paths), use-forbidden-name
     (single-letter loop indices are Solidity convention).

### 7. Add SOLHINT.md
Per-rule rationale for every disabled rule, so future contributors
can understand why each was turned off and re-enable on a per-line
basis if a rule is appropriate for new code.

## Outcome

| Metric                | Before  | After |
|-----------------------|---------|-------|
| solhint problems      | 3,746   | 0     |
| solhint errors        | 0       | 0     |
| no-global-import      | 95      | 0     |
| no-unused-import      | 0       | 0     |
| avoid-low-level-calls | 6       | 0     |
| no-unused-vars        | 6       | 0     |
| explicit-types        | 3       | 0     |

End-to-end across waves 1-3:
  - Wave 1: 13,414 -> 13,003 (gas-increment-by-one: 433 -> 22)
  - Wave 2: 13,003 -> 3,746  (use-natspec: 9,257 -> 0)
  - Wave 3: 3,746 -> 0       (named-imports + config policy)

All 94 contract files parse cleanly with @solidity-parser/parser.
Stacks on top of #250 (wave 2).

+++ contracts/RevenueSplitter.sol
@@ -53,7 +53,7 @@ contract RevenueSplitter is ReentrancyGuard {
         owner = msg.sender;
         
         uint256 length = _accounts.length;
-        for (uint i = 0; i < length; ++i) {
+        for (uint256 i = 0; i < length; ++i) {
             require(_accounts[i] != address(0), "zero address");
             require(_shares[i] > 0, "zero share");
             payees.push(Payee({account: _accounts[i], shareBps: _shares[i]}));
@@ -76,7 +76,7 @@ contract RevenueSplitter is ReentrancyGuard {
         uint256 payeesFailed = 0;
 
         uint256 length = payees.length;
-        for (uint i = 0; i < length; ++i) {
+        for (uint256 i = 0; i < length; ++i) {
             uint256 amount;
             if (i == length - 1) {
                 amount = balance - distributed; // Give remainder to last
@@ -88,6 +88,7 @@ contract RevenueSplitter is ReentrancyGuard {
                 // H-29 FIX: Compute amount for last payee BEFORE updating distributed.
                 // Only increment distributed after a successful transfer.
                 // M-2 FIX: Low-level call for non-standard ERC20s (USDT)
+                // solhint-disable-next-line avoid-low-level-calls
                 (bool callOk, bytes memory returnData) = token.call(
                     abi.encodeWithSelector(IERC20.transfer.selector, payees[i].account, amount)
                 );
